Updated PCB to v1 for bulk order
Moved a trace on the bottom layer to the edge to avoid splitting the ground plane, it travels on the upper layer to get to and from the edge and then jumps to the bottom layer to cross the lines in the way.
Additionally, vertical bars have been added next to all pins that need to be connected on the surface mount pads in order to mount the RaspberryPi Pico with its castellated edges as a surface mount component.